Added /client/apps to the executable path.
This required changes to both the Dispatch and TabCompletion classes so
that they each are constructed with a collection of collections, to allow
some to be static and others dynamic. The dynamic component in each
instance is the new AppsInPath class, which enumerates files in a given
directory that have the .app extension.